Removed direct reference to WebGLRenderingContext
We need a headless browser for testing purposes, which natually won't provide
any WebGL functionality. In the futrue we may think about omse off screen rendering
setup for testing GL features but for now we don't want to directly reference
webgl with any symbols in our code. The only area that did that was the constants
to keyword mapping, which no gets the constants from a context instance rather than
looking them up statically.